channel utility: fix bug with signed int

+++ elements/brn2/wifi/ath/ath2decap.cc
@@ -150,7 +150,7 @@ Ath2Decap::simple_action(Packet *p)
     if ( eh->retries < ath2_h->anno.tx.ts_longretry )
       eh->retries = ath2_h->anno.tx.ts_longretry;
 
-    if ( (_cst != NULL) && (ath2_h->anno.tx.ts_channel_utility != CHANNEL_UTILITY_INVALID) )
+    if ( (_cst != NULL) && ((uint8_t)ath2_h->anno.tx.ts_channel_utility != CHANNEL_UTILITY_INVALID) )
       _cst->addHWStat(&(p->timestamp_anno()), ath2_h->anno.tx.ts_channel_utility, 0, 0);
 
     BRNPacketAnno::set_channel_anno(q, ath2_h->anno.tx.ts_channel); 
@@ -205,7 +205,7 @@ Ath2Decap::simple_action(Packet *p)
 
     BRNPacketAnno::set_channel_anno(q, ath2_h->anno.rx.rs_channel);
 
-    if ( (_cst != NULL) && (ath2_h->anno.rx.rs_channel_utility != CHANNEL_UTILITY_INVALID) )
+    if ( (_cst != NULL) && ((uint8_t)ath2_h->anno.rx.rs_channel_utility != CHANNEL_UTILITY_INVALID) )
       _cst->addHWStat(&(p->timestamp_anno()), ath2_h->anno.rx.rs_channel_utility, 0, 0);
   }
